Move ClearCommand to global
