+++ src/main/java/seedu/address/logic/commands/ClearCommand.java
@@ -1,10 +1,8 @@
-package seedu.address.logic.commands.contacts;
+package seedu.address.logic.commands;
 
 import static java.util.Objects.requireNonNull;
 
 import seedu.address.logic.CommandHistory;
-import seedu.address.logic.commands.Command;
-import seedu.address.logic.commands.CommandResult;
 import seedu.address.model.AddressBook;
 import seedu.address.model.Model;
