+++ src/test/java/seedu/address/logic/commands/ClearCommandTest.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.commands.contacts;
+package seedu.address.logic.commands;
 
 import static seedu.address.logic.commands.CommandTestUtil.assertCommandSuccess;
 import static seedu.address.testutil.TypicalAddressBook.getTypicalAddressBook;
