+++ src/test/java/systemtests/contacts/AddCommandSystemTest.java
@@ -87,7 +87,7 @@ public class AddCommandSystemTest extends ContactsSystemTest {
         assertCommandSuccess(command, toAdd);
 
         /* Case: add to empty address book -> added */
-        deleteAllPersons();
+        clear();
         assertCommandSuccess(ALICE);
 
         /* Case: add a person with tags, command with parameters in random order -> added */
